Add some definitions
This library has lagged behind a number of the others, so this moves over some
of the existing definitions. This doesn’t affect any existing definitions (even
in places where the current naming is out of sync with other libraries), so it’s
a compatible change … except that it also adds a dependency on the comonad
package.